requirements.txt
fastapi==0.72.0
uvicorn==0.17.0
python-multipart==0.0.5
jinja2==3.0.3
pydantic==1.9.0
sqlalchemy==1.4.29
pymysql==1.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0